config(ai-app): 重构配置文件结构并实现网关访问频率限制
- 将 application.yaml 中的 profiles 从 data,middleware,ollama 改为 dev
- 移除 application.yaml 中的 server 和 logging 配置
- 重命名 application-data.yaml 为 application-dev.yaml
- 在 application-dev.yaml 中新增 Nacos 配置、AI 相关配置、Gorse 配置和服务器端口日志设置
- 删除独立的 application-ollama.yaml 配置文件
- 重命名 application-middleware.yaml 为 application-prod.yaml
- 在 application-prod.yaml 中新增完整的生产环境配置包括数据库、Redis、MongoDB等
- 修改 EmbeddingProductServiceImpl 中的向量搜索参数计算方式
- 调整行为服务和商品服务的配置文件结构，统一配置管理方式
- 重命名 gateway 配置文件，将 application-data.yaml 改为 application-dev.yaml
- 在网关过滤器中实现基于 Redis 的 IP 访问频率限制功能
- 新增 IP 访问统计的数据结构和计数逻辑
- 移除网关中多余的注释代码并优化代码结构
- 清理 goods-app 中的测试代码文件 GoodsAppApplicationTest.java

+++ gsy-ai-cloud/ai-server/src/main/java/com/guanshiyun/service/embedding/impl/EmbeddingProductServiceImpl.java
@@ -148,8 +148,8 @@ public class EmbeddingProductServiceImpl implements EmbeddingProductService {
 
             // 并行执行向量搜索和 Gorse 推荐
             return Mono.zip(
-                            vectorSearch(fusedQuery, (int)(topK * 2)),
-                            gorseClient.getRecommend(userId.toString(), (int)(topK * 2))
+                            vectorSearch(fusedQuery, topK * 2),
+                            gorseClient.getRecommend(userId.toString(), topK * 2)
                                     .map(ids -> ids.stream()
                                             .map(myLong::myLong)
                                             .toList())
